clean up the broken-apart pieces of the srl_bot add ability to use a debugging channel when testing locally
