+++ alttprbot/util/srl.py
@@ -1,6 +1,7 @@
 from config import Config as c
 import aiofiles
 import json
+import re
 from . import http
 
 async def get_race(raceid, complete=False):
@@ -19,6 +20,15 @@ async def get_race(raceid, complete=False):
             return {}
 
     return await http.request_generic(f'http://api.speedrunslive.com/races/{raceid}', returntype='json')
+
+def srl_race_id(channel):
+    if channel == '#srl-synack-testing':
+        return 'test'
+    if re.search('^#srl-[a-z0-9]{5}$', channel):
+        return channel.partition('-')[-1]
+
+async def get_all_races():
+    return await http.request_generic(f'http://api.speedrunslive.com/races', returntype='json')
 
 async def send_irc_message(raceid, message):
     if c.DEBUG: return
