+++ alttprbot_srl/commands.py
@@ -1,6 +1,192 @@
 import argparse
 import sys
 import shlex
+from alttprbot.util.srl import get_all_races, get_race, srl_race_id
+from alttprbot.database import srl_races
+from alttprbot.alttprgen import preset, mystery, spoilers
+from alttprbot.database import spoiler_races, srl_races
+from alttprbot.smz3gen import preset as smz3_preset
+from alttprbot.smz3gen import spoilers as smz3_spoilers
+
+
+async def handler(target, source, message, client):
+    if not message[0] == '$': return
+
+    try:
+        args = parse_args(message)
+    except argparse.ArgumentError as e:
+        if not target == '#speedrunslive':
+            await client.message(target, e.message)
+        return
+
+    if args.command == '$preset' and target.startswith('#srl-'):
+        srl_id = srl_race_id(target)
+        srl = await get_race(srl_id)
+        await client.message(target, "Generating game, please wait.  If nothing happens after a minute, contact Synack.")
+        race = await srl_races.get_srl_race_by_id(srl_id)
+
+        if race:
+            await client.message(target, "There is already a game generated for this room.  To cancel it, use the $cancel command.")
+            return
+
+        if srl['game']['abbrev'] == 'alttphacks':
+            try:
+                seed, preset_dict = await preset.get_preset(args.preset, hints=args.hints, spoilers="off")
+            except preset.PresetNotFoundException:
+                await client.message(target, "That preset does not exist.  For documentation on using this bot, visit https://sahasrahbot.synack.live")
+
+            goal_name = preset_dict['goal_name']
+            goal = f"vt8 randomizer - {goal_name}"
+            code = await seed.code()
+            if args.silent:
+                await client.message(target, f"{goal} - {seed.url} - ({'/'.join(code)})")
+            else:
+                await client.message(target, f".setgoal {goal} - {seed.url} - ({'/'.join(code)})")
+        elif srl['game']['abbrev'] == 'alttpsm':
+            seed = await smz3_preset.get_preset(args.preset)
+            goal = 'beat the games'
+            if args.silent:
+                await client.message(target, f"{goal} - {args.preset} - {seed.url}")
+            else:
+                await client.message(target, f".setgoal {goal} - {args.preset} - {seed.url}")
+        else:
+            await client.message(target, "This game is not yet supported.")
+            return
+
+        await srl_races.insert_srl_race(srl_id, goal)
+
+    if args.command == '$random' and target.startswith('#srl-'):
+        srl_id = srl_race_id(target)
+        srl = await get_race(srl_id)
+        await client.message(target, "Generating game, please wait.  If nothing happens after a minute, contact Synack.")
+        race = await srl_races.get_srl_race_by_id(srl_id)
+
+        if race:
+            await client.message(target, "There is already a game generated for this room.  To cancel it, use the $cancel command.")
+            return
+
+        if srl['game']['abbrev'] == 'alttphacks':
+            try:
+                seed = await mystery.generate_random_game(weightset=args.weightset, tournament=True, spoilers="off")
+            except mystery.WeightsetNotFoundException:
+                await client.message(target, "That weightset does not exist.  For documentation on using this bot, visit https://sahasrahbot.synack.live")
+                return
+
+            code = await seed.code()
+            goal = f"vt8 randomizer - random {args.weightset}"
+            if args.silent:
+                await client.message(target, f"{goal} - {seed.url} - ({'/'.join(code)})")
+            else:
+                await client.message(target, f".setgoal {goal} - {seed.url} - ({'/'.join(code)})")
+        else:
+            await client.message(target, "This game is not yet supported.")
+            return
+
+        await srl_races.insert_srl_race(srl_id, goal)
+
+    if args.command == '$mystery' and target.startswith('#srl-'):
+        srl_id = srl_race_id(target)
+        srl = await get_race(srl_id)
+        await client.message(target, "Generating game, please wait.  If nothing happens after a minute, contact Synack.")
+        race = await srl_races.get_srl_race_by_id(srl_id)
+
+        if race:
+            await client.message(target, "There is already a game generated for this room.  To cancel it, use the $cancel command.")
+            return
+
+        if srl['game']['abbrev'] == 'alttphacks':
+            try:
+                seed = await mystery.generate_random_game(weightset=args.weightset, tournament=True, spoilers="mystery")
+            except mystery.WeightsetNotFoundException:
+                await client.message(target, "That weightset does not exist.  For documentation on using this bot, visit https://sahasrahbot.synack.live")
+                return
+
+            code = await seed.code()
+            goal = f"vt8 randomizer - mystery {args.weightset}"
+            if args.silent:
+                await client.message(target, f"{goal} - {seed.url} - ({'/'.join(code)})")
+            else:
+                await client.message(target, f".setgoal {goal} - {seed.url} - ({'/'.join(code)})")
+        else:
+            await client.message(target, "This game is not yet supported.")
+            return
+
+        await srl_races.insert_srl_race(srl_id, goal)
+
+    if args.command == '$custom' and target.startswith('#srl-'):
+        await client.message(target, "Not yet implemented.  Sorry!")
+
+    if args.command == '$spoiler' and target.startswith('#srl-'):
+        srl_id = srl_race_id(target)
+        srl = await get_race(srl_id)
+        await client.message(target, "Generating game, please wait.  If nothing happens after a minute, contact Synack.")
+        race = await srl_races.get_srl_race_by_id(srl_id)
+
+        if race:
+            await client.message(target, "There is already a game generated for this room.  To cancel it, use the $cancel command.")
+            return
+
+        if srl['game']['abbrev'] == 'alttphacks':
+            try:
+                seed, preset_dict, spoiler_log_url = await spoilers.generate_spoiler_game(args.preset)
+            except preset.PresetNotFoundException:
+                await client.message(target, "That preset does not exist.  For documentation on using this bot, visit https://sahasrahbot.synack.live")
+
+            goal_name = preset_dict['goal_name']
+
+            if not seed:
+                return
+
+            goal = f"vt8 randomizer - spoiler {goal_name}"
+            studytime = 900 if not args.studytime else args.studytime 
+            code = await seed.code()
+            if args.silent:
+                await client.message(target, f"{goal} - {seed.url} - ({'/'.join(code)})")
+            else:
+                await client.message(target, f".setgoal {goal} - {seed.url} - ({'/'.join(code)})")
+            await client.message(target, f"The spoiler log for this race will be sent after the race begins in SRL.  A {studytime}s countdown timer at that time will begin.")
+        elif srl['game']['abbrev'] == 'alttpsm':
+            seed, spoiler_log_url = await smz3_spoilers.generate_spoiler_game(args.preset)
+
+            if seed is None:
+                await client.message(target, "That preset does not exist.  For documentation on using this bot, visit https://sahasrahbot.synack.live")
+                return
+
+            goal = f"spoiler beat the games"
+            studytime = 1500 if not args.studytime else args.studytime 
+            if args.silent:
+                await client.message(target, f"{goal} - {seed.url}")
+            else:
+                await client.message(target, f".setgoal {goal} - {seed.url}")
+            await client.message(target, f"The spoiler log for this race will be sent after the race begins in SRL.  A {studytime}s countdown timer at that time will begin.")
+        else:
+            await client.message(target, "This game is not yet supported.")
+            return
+
+        await srl_races.insert_srl_race(srl_id, goal)
+        await spoiler_races.insert_spoiler_race(srl_id, spoiler_log_url, studytime)
+
+    if args.command == '$cancel' and target.startswith('#srl-'):
+        srl_id = srl_race_id(target)
+        await srl_races.delete_srl_race(srl_id)
+        await spoiler_races.delete_spoiler_race(srl_id)
+        await client.message(target, "Current race cancelled.")
+        await client.message(target, f".setgoal new race")
+
+    if args.command == '$help':
+        await client.message(target, "For documentation on using this bot, visit https://sahasrahbot.synack.live")
+
+    if args.command == '$joinroom':
+        await client.join(args.channel)
+
+    if args.command == '$leave' and target.startswith('#srl-'):
+        await client.part(target)
+
+    if args.command == '$vt' and target.startswith('#srl-'):
+        await client.message(target, "You summon VT, he looks around confused and curses your next game with bad RNG.")
+
+    if args.command == '$echo':
+        await client.message(source, args.message)
 
 def parse_args(message):
     split_msg = ['sb'] + shlex.split(message)
